// packages/quyx/shared/types.ts
export type HTMLTag =
    | {
        tag: 'title'
        attrs?: Record<string, any>
        children: string
    }
    | {
        tag: 'meta' | 'link'
        attrs: Record<string, any>
        children?: never
    }
    | {
        tag: 'script'
        attrs: Record<string, any>
        children: string
    }
    | {
        tag: 'style'
        attrs: Record<string, any>
        children: string
    }